requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
moviepy==1.0.3
pillow==10.3.0
requests==2.32.3
pysrt==1.1.2
boto3==1.34.144
pydantic==2.8.2
